abstract output
